backend/requirements.txt
# Django requirements
django
djangorestframework
pandas
openpyxl
django-cors-headers
reportlab
